Add @psalm-pure annotations
`@psalm-pure` tells the static analyzer psalm that the output of a function is only dependent on the input parameters (the output will always be the same for the same input). Adding this to the library helps psalm further analyze code and to add `@psalm-pure` to applications where the `preg::` functions are used.